+++ apps/aspeed-pfr/src/cerberus_pfr/cerberus_pfr_key_cancellation.h
@@ -9,6 +9,6 @@
 #if defined(CONFIG_CERBERUS_PFR)
 #define KEY_CANCELLATION_MAX_KEY_ID 127
 
-int verify_csk_key_id(struct pfr_manifest *manifest, uint32_t key_id);
-int cancel_csk_key_id(struct pfr_manifest *manifest, uint32_t key_id);
+int verify_csk_key_id(struct pfr_manifest *manifest, uint8_t key_id);
+int cancel_csk_key_id(struct pfr_manifest *manifest, uint8_t key_id);
 #endif
